+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.1.2 ()
+* add support of `compact` option for toml
+
 # 0.1.1 (2022-06-28)
 
 * add more tests
@@ -6,4 +9,4 @@
 # 0.1.0 (2022-06-28)
 
 * initial version with support of `json`, `yaml` and `toml`
-* json supports compact option
+* json supports `compact` option